new feature: polygon datatype to smw
Ok, so what i have done so far is to add the necessary changes to allow
for a polygon datatype which is a container. also added nuts and bolts where
it seemed to be necessary in order to make the cogwheels spin.

The datastructure seems good, and when i try to store data i can see that it
gets persisted to the database. However, it doesn't seem to handle reading the
data through #ask afterwards, any ideas on this?.

Change-Id: I068e9212adfd04b452d2e6b00e9d198537b2761b